+++ pub/unrouter/lib/src/shell/shell.dart
@@ -1,1 +1,2 @@
-export 'shell_runtime_binding.dart';
+export 'shell_coordinator.dart';
+export 'shell_route_binding.dart';